requirements.txt
crewai
crewai_tools
langchain
Streamlit
yt_dlp
youtube-transcript-api
load_dotenv